+++ runtests.py
@@ -26,6 +26,15 @@ installed_apps = [
     "django.contrib.messages",
 ]
 
+
+class DisableMigrations:
+    def __contains__(self, item):
+        return True
+
+    def __getitem__(self, item):
+        return None
+
+
 DEFAULT_SETTINGS = dict(
     ALLOWED_HOSTS=["localhost"],
     AUTH_USER_MODEL="custom_user.CustomUser",
@@ -37,18 +46,17 @@ DEFAULT_SETTINGS = dict(
         "default": {"ENGINE": "django.db.backends.sqlite3"},
         "other": {"ENGINE": "django.db.backends.sqlite3"},
     },
-    TEMPLATES=[
-        {
-            "BACKEND": "django.template.backends.django.DjangoTemplates",
-            "APP_DIRS": True,
-            "OPTIONS": {
-                "context_processors": [
-                    "django.contrib.auth.context_processors.auth",
-                    "django.contrib.messages.context_processors.messages",
-                ]
-            },
-        }
-    ],
+    MIGRATION_MODULES=DisableMigrations(),
+    TEMPLATES=[{
+        'BACKEND': 'django.template.backends.django.DjangoTemplates',
+        'APP_DIRS': True,
+        'OPTIONS': {
+            'context_processors': [
+                'django.contrib.auth.context_processors.auth',
+                'django.contrib.messages.context_processors.messages',
+            ]
+        },
+    }],
 )
 MIDDLEWARE = [
     "django.contrib.sessions.middleware.SessionMiddleware",
@@ -67,7 +75,8 @@ def main():
     if not settings.configured:
         settings.configure(**DEFAULT_SETTINGS)
     django.setup()
-    failures = DiscoverRunner(failfast=False).run_tests(["simple_history.tests"])
+    failures = DiscoverRunner(failfast=False).run_tests(
+        ["simple_history.tests"])
     failures |= DiscoverRunner(failfast=False).run_tests(
         ["simple_history.registry_tests"]
     )